requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.8
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
openai==0.27.8
python-dotenv==1.0.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sniffio==1.3.1
tqdm==4.66.5
urllib3==2.2.2
yarl==1.9.4
